Fix: signed/unsigned mismatch in potion_dip.

+++ src/potion.c
@@ -3169,7 +3169,7 @@ potion_dip(struct obj *obj, struct obj *potion)
             pline1(nothing_happens);
         } else {
             short save_otyp = obj->otyp;
-            int save_dknown = (int) obj->dknown;
+            unsigned int save_dknown = obj->dknown;
 
             /* KMH, conduct */
             if (!u.uconduct.polypiles++)
@@ -3188,7 +3188,7 @@ potion_dip(struct obj *obj, struct obj *potion)
                     makeknown(POT_POLYMORPH);
                 return ECMD_TIME;
             } else if (obj->otyp != save_otyp
-                    || obj->dknown != (int) save_dknown) {
+                    || obj->dknown != save_dknown) {
                 if (potion->dknown)
                     makeknown(POT_POLYMORPH);
                 useup(potion);
